Add entire ARM instruction set and a basic disassembler for ARM

+++ gba/gamepak.c
@@ -3,7 +3,7 @@
 #include <stdlib.h>
 #include <gba/gamepak.h>
 
-bool initGamePak(GamePak* gamepak, uint8_t* allocated, long size) {
+bool initGamePak(GamePak* gamepak, uint8_t* allocated, size_t size) {
 	gamepak->allocated = allocated;
 	gamepak->inserted = false;
 	gamepak->size = (size_t)size;
